Update md.ini
Changed Sumtu to Cumtu

As this language was still developing, the name was not consistant. However, now the Cumtu Literacy Committee decided to use "Cumtu" instead of "Sumtu" as the "C" sound is more similar to local language pronunciation whilst "S" sound like a bit deeper which is not matching with local's pronunciation.